+++ src/com/affixus/pojo/Order.java
@@ -22,7 +22,7 @@ public class Order extends BasePojo {
 	private Process rm;
 	private Process cast;
 	private String t_charges;
-	private float camGrams;
+	private double camGrams;
 	
 	public String get_id() {
 		return _id;
@@ -108,10 +108,10 @@ public class Order extends BasePojo {
 	public void setT_charges(String t_charges) {
 		this.t_charges = t_charges;
 	}
-	public float getCamGrams() {
+	public double getCamGrams() {
 		return camGrams;
 	}
-	public void setCamGrams(float camGrams) {
+	public void setCamGrams(double camGrams) {
 		this.camGrams = camGrams;
 	}
